+++ src/AngularSPAWebAPI/app/account/signin/signin.component.ts
@@ -1,15 +1,17 @@
 ﻿import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 
-import { AuthenticationService } from '../services/authentication.service';
-import { Signin } from './signin';
+import { AuthenticationService } from '../../services/authentication.service';
+import { Signin } from '../signin';
 
 @Component({
-    templateUrl: 'signin.component.html'
+    templateUrl: './signin.component.html'
 })
 export class SigninComponent extends Signin {
 
-    constructor(public router: Router, public authenticationService: AuthenticationService) {
+    constructor(
+        protected router: Router,
+        protected authenticationService: AuthenticationService) {
         super(router, authenticationService);
 
         // Preloads data for live example.