+++ src/AngularSPAWebAPI/app/account/signup/signup.component.ts
@@ -1,25 +1,24 @@
 ﻿import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 
-import { AuthenticationService } from '../services/authentication.service';
-import { IdentityService } from '../services/identity.service';
-import { Signin } from './signin';
+import { AuthenticationService } from '../../services/authentication.service';
+import { IdentityService } from '../../services/identity.service';
+import { Signin } from '../signin';
 
 @Component({
-    templateUrl: 'signup.component.html'
+    templateUrl: './signup.component.html'
 })
 export class SignupComponent extends Signin {
 
     constructor(
-        public router: Router,
-        public authenticationService: AuthenticationService,
-        private identityService: IdentityService
-    ) {
+        protected router: Router,
+        protected authenticationService: AuthenticationService,
+        private identityService: IdentityService) {
         super(router, authenticationService);
     }
 
     signup(): void {
-        this.identityService.Create(this.model)
+        this.identityService.create(this.model)
             .subscribe(
             (res: any) => {
                 // IdentityResult.