+++ src/AngularSPAWebAPI/app/app-routing.module.ts
@@ -1,29 +1,26 @@
 ﻿import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-
-import { AuthGuard } from './services/auth.guard';
+import { Routes, RouterModule, PreloadAllModules } from '@angular/router';
 
 import { HomeComponent } from './home/home.component';
-import { ResourcesComponent } from './resources/resources.component';
-import { DashboardComponent } from './dashboard/dashboard.component';
-import { SigninComponent } from './account/signin.component';
-import { SignupComponent } from './account/signup.component';
 
-// We use PathLocationStrategy - the default "HTML 5 pushState" style.
-// https://angular.io/docs/ts/latest/guide/router.html#!#browser-url-styles
-// Router on the server (see Startup.cs) must match the router on the client to use PathLocationStrategy.
+// We use PathLocationStrategy - the default "HTML 5 pushState" style:
+// - https://angular.io/docs/ts/latest/guide/router.html#!#browser-url-styles
+// - Router on the server (see Startup.cs) must match the router on the client to use PathLocationStrategy
+// and Lazy Loading Modules:
+// - https://angular.io/guide/ngmodule#lazy-loading-modules-with-the-router
 const routes: Routes = [
     { path: '', redirectTo: 'home', pathMatch: 'full' },
     { path: 'home', component: HomeComponent },
-    { path: 'resources', component: ResourcesComponent, canActivate: [AuthGuard] },
-    { path: 'dashboard', component: DashboardComponent, canActivate: [AuthGuard] },
-    { path: 'signin', component: SigninComponent },
-    { path: 'signup', component: SignupComponent }
+    { path: 'account', loadChildren: './account/account.module#AccountModule' },
+    { path: 'resources', loadChildren: './resources/resources.module#ResourcesModule' },
+    { path: 'dashboard', loadChildren: './dashboard/dashboard.module#DashboardModule' }
 ];
 
 @NgModule({
     imports: [
-        RouterModule.forRoot(routes)
+        RouterModule.forRoot(routes, {
+            preloadingStrategy: PreloadAllModules
+        })
     ],
     exports: [RouterModule]
 })