+++ src/AngularSPAWebAPI/app/app.component.ts
@@ -1,32 +1,45 @@
-﻿import { Component } from '@angular/core';
+﻿import { Component, OnInit } from '@angular/core';
+import { Title } from '@angular/platform-browser';
 import { Router } from '@angular/router';
 import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/map';
 
 import { AuthenticationService } from './services/authentication.service';
+import { User } from './models/user';
 
 @Component({
     selector: 'app-component',
-    templateUrl: 'app.component.html'
+    templateUrl: './app.component.html'
 })
-export class AppComponent {
+export class AppComponent implements OnInit {
+
+    navItems: any[] = [
+        { name: 'Home', route: 'home' },
+        { name: 'Resources', route: 'resources' }
+    ];
 
     signedIn: Observable<boolean>;
 
-    name: Observable<string>;
+    name: string;
+    isAdmin: boolean;
 
-    isAdmin: Observable<boolean>;
+    constructor(
+        public title: Title,
+        private authenticationService: AuthenticationService,
+        private router: Router
+    ) { }
 
-    constructor(public authenticationService: AuthenticationService, private router: Router) {
+    ngOnInit() {
+        this.title.setTitle('Angular SPA WebAPI');
+
         this.signedIn = this.authenticationService.isSignedIn();
 
-        this.name = this.authenticationService.getUser()
-            .map((user: any) => (typeof user.given_name !== "undefined") ? user.given_name : null);
+        this.authenticationService.userChanged().subscribe(
+            (user: User) => {
+                this.name = user.givenName;
+                this.isAdmin = this.authenticationService.isInRole("administrator");
+            });
 
-        this.isAdmin = this.authenticationService.getRoles()
-            .map((roles: string[]) => roles.indexOf("administrator") != -1);
-
-        // Optional strategy for refresh token through a scheduler.
+        // Strategy for refresh token through a scheduler.
         this.authenticationService.startupTokenRefresh();
     }