+++ src/AngularSPAWebAPI/app/main-aot.ts
@@ -1,15 +1,15 @@
 // AoT compilation entry point.
 
-// Polyfills (no need reflect-metadata).
+// Polyfills.
 import 'core-js/client/shim';
 import 'zone.js/dist/zone';
 
+import 'web-animations-js';
+
 import 'hammerjs';
 
-// Angular Material theme.
-import './styles/blue-amber.scss';
-// App styles.
-import './styles/app.scss';
+// Styles.
+import './styles.scss';
 
 import { platformBrowser } from '@angular/platform-browser';
 import { enableProdMode } from '@angular/core';