+++ src/AngularSPAWebAPI/app/main.ts
@@ -2,15 +2,14 @@
 
 // Polyfills.
 import 'core-js/client/shim';
-import 'reflect-metadata';
 import 'zone.js/dist/zone';
+
+import 'web-animations-js';
 
 import 'hammerjs';
 
-// Angular Material theme.
-import './styles/blue-amber.scss';
-// App styles.
-import './styles/app.scss';
+// Styles.
+import './styles.scss';
 
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 import { AppModule } from './app.module';