+++ src/AngularSPAWebAPI/app/resources/resources.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+
+import { ResourcesRoutingModule } from './resources-routing.module';
+import { SharedModule } from '../shared/shared.module';
+
+import { ResourcesComponent } from './resources.component';
+
+@NgModule({
+    imports: [
+        ResourcesRoutingModule,
+        SharedModule
+    ],
+    declarations: [
+        ResourcesComponent
+    ]
+})
+export class ResourcesModule { }